Backend/requirements.txt
asgiref==3.8.1
Django==5.2.3
django-cors-headers==4.7.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
pillow==11.2.1
psycopg2==2.9.10
PyJWT==2.9.0
pyotp==2.9.0
sqlparse==0.5.3
tzdata==2025.2
